+++ wait_for2/impl.py
@@ -9,7 +9,7 @@ from asyncio import CancelledError, ensure_future, TimeoutError
 
 try:
     from asyncio import get_running_loop
-except ImportError:
+except ImportError:  # pragma: no cover
     from asyncio import get_event_loop as get_running_loop
 from functools import partial
 
@@ -19,24 +19,75 @@ def _release_waiter(waiter, *args):  # copied from from asyncio.tasks
         waiter.set_result(None)
 
 
-async def _cancel_and_wait(fut, loop):  # copied from from asyncio.tasks
-    waiter = loop.create_future()
-    cb = partial(_release_waiter, waiter)
-    fut.add_done_callback(cb)
-    try:
+def _handle_cancelling_with_inner_completion(loop, fut, fut_result, res_exception, race_handler):
+    if race_handler:
+        try:
+            race_handler(fut_result, res_exception)
+        except Exception as e:
+            loop.call_exception_handler({"message": "wait_for2 race_handler failed", "exception": e, "future": fut})
+    raise CancelledWithResultError(fut_result, res_exception)
+
+
+async def _cancel_and_wait2(fut, loop, cancelling, race_handler):
+    """
+    The builtin implementation of _cancel_and_wait is made in a way to ensure cancellation of it will always be
+    possible, at the cost of ensuring that the wrapped future shall terminate inside it.
+
+    The documentation says that
+        - "If a timeout occurs, it cancels the task and raises asyncio.TimeoutError."
+        - "The function will wait until the future is actually cancelled"
+    , but their combination will result in undocumented behaviour:
+        If a timeout occurs and the inner future is cancelled, then cancelling the waiter will instantly cancel
+        it and the inner future may still be running, which does not fulfill the second promise.
+
+    This implementation will prioritize cancellation or the result of the inner future dynamically as it makes sense.
+    """
+    if not cancelling:
+        # We need to detect explicit cancellations so the good-case value returning will not be used.
+        waiter = loop.create_future()
+        cb = partial(_release_waiter, waiter)
+        fut.add_done_callback(cb)
         fut.cancel()
-        await waiter
-    finally:
-        fut.remove_done_callback(cb)
+        try:
+            await waiter
+        except CancelledError:
+            cancelling = True  # explicitly cancelling the inner from now on
+        finally:
+            fut.remove_done_callback(cb)
+    else:
+        fut.cancel()
+    # At this point there's no benefit of wrapping the future with a waiter since we're cancelling it?
+    try:
+        fut_result = await fut
+        if not cancelling:
+            return fut_result
+    except CancelledError as exc:
+        if cancelling:
+            raise exc
+        raise TimeoutError() from exc
+    except Exception as exc:
+        if not cancelling:
+            raise exc
+        fut_result = exc
+        res_exception = True
+    else:
+        res_exception = False
+    # The waiting construct is already being cancelled, we need to discard/handle the inner future's
+    # result here to adhere to the cancellation request.
+    _handle_cancelling_with_inner_completion(loop, fut, fut_result, res_exception, race_handler)
 
 
 class CancelledWithResultError(CancelledError):
-    def __init__(self, result):
-        super(CancelledWithResultError, self).__init__(result)
+    def __init__(self, result, exc):
+        super(CancelledWithResultError, self).__init__(result, exc)
 
     @property
     def result(self):
         return self.args[0]
+
+    @property
+    def is_exception(self):
+        return self.args[1]
 
 
 async def wait_for(fut, timeout, *, loop=None, race_handler=None):
@@ -63,13 +114,20 @@ async def wait_for(fut, timeout, *, loop=None, race_handler=None):
     It will be called with the result of the future when the waiter task is being cancelled. Even if this is provided,
     the special error will be raised in the place of a normal CancelledError.
 
+    Additionally, this implementation will inherit the behaviour of the inner future when it comes to ignoring
+    cancellation. The builtin version prefers to always be cancellable, even if that means the wrapped future may
+    not be terminated with it. (behaviour of builtin _cancel_and_wait) This behaviour is also improved in
+    timeout-cancel edge cases, where the builtin would not wait for the termination of the inner future if the
+    waiter was cancelled after timeout handling had already started. This is more consistent as the inner future
+    must always be stopped for it to return.
+
     NOTE: CancelledWithResultError is limited to the coroutine wait_for is invoked from!
     If this wait_for is wrapped in tasks those will not propagate the special exception, but raise their own
     CancelledError instances.
     """
     if loop is None:
         loop = get_running_loop()
-    elif sys.version_info >= (3, 10):
+    elif sys.version_info >= (3, 10):  # pragma: no cover
         raise RuntimeError("loop parameter has been dropped since Python 3.10")
 
     if timeout is None:
@@ -81,13 +139,7 @@ async def wait_for(fut, timeout, *, loop=None, race_handler=None):
         if fut.done():
             return fut.result()
 
-        await _cancel_and_wait(fut, loop=loop)
-        try:
-            fut.result()
-        except CancelledError as exc:
-            raise TimeoutError() from exc
-        else:
-            raise TimeoutError()
+        return await _cancel_and_wait2(fut, loop, False, race_handler)
 
     waiter = loop.create_future()
     timeout_handle = loop.call_later(timeout, _release_waiter, waiter)
@@ -102,27 +154,21 @@ async def wait_for(fut, timeout, *, loop=None, race_handler=None):
             if fut.done():
                 try:
                     fut_result = fut.exception()
-                    if fut_result is None:
-                        fut_result = fut.result()
                 except CancelledError:
                     raise  # inner future was also cancelled
-                if race_handler:
-                    try:
-                        race_handler(fut_result)
-                    except Exception as e:
-                        loop.call_exception_handler(
-                            {"message": "wait_for2 race_handler failed", "exception": e, "future": fut}
-                        )
-                raise CancelledWithResultError(fut_result)
+                if fut_result is None:
+                    fut_result = fut.result()
+                    res_exception = False
+                else:
+                    res_exception = True
+                _handle_cancelling_with_inner_completion(loop, fut, fut_result, res_exception, race_handler)
             fut.remove_done_callback(cb)
-            await _cancel_and_wait(fut, loop=loop)
-            raise
+            await _cancel_and_wait2(fut, loop, True, race_handler)
 
         if fut.done():
             return fut.result()
         else:
             fut.remove_done_callback(cb)
-            await _cancel_and_wait(fut, loop=loop)
-            raise TimeoutError()
+            return await _cancel_and_wait2(fut, loop, False, race_handler)
     finally:
         timeout_handle.cancel()